+++ pkg/compactor/bucket_compactor_test.go
@@ -172,7 +172,7 @@ func TestFilterOwnJobs(t *testing.T) {
 	m := NewBucketCompactorMetrics(prometheus.NewCounter(prometheus.CounterOpts{}), prometheus.NewCounter(prometheus.CounterOpts{}), nil)
 	for testName, testCase := range tests {
 		t.Run(testName, func(t *testing.T) {
-			bc, err := NewBucketCompactor(log.NewNopLogger(), nil, nil, nil, nil, "", nil, 2, false, testCase.ownJob, m)
+			bc, err := NewBucketCompactor(log.NewNopLogger(), nil, nil, nil, nil, "", nil, 2, false, testCase.ownJob, nil, m)
 			require.NoError(t, err)
 
 			res, err := bc.filterOwnJobs(jobsFn())